refactor: move InputValidator to view package
